formatting and adding routes
